# Add the possibility to provide “custom” metric-store to the builder