--- a/internal/store/deployment.go
+++ b/internal/store/deployment.go
@@ -18,6 +18,7 @@ package store
 
 import (
 	"k8s.io/kube-state-metrics/pkg/metric"
+	generator "k8s.io/kube-state-metrics/pkg/metric_generator"
 
 	v1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -33,7 +34,7 @@ var (
 	descDeploymentLabelsHelp          = "Kubernetes labels converted to Prometheus labels."
 	descDeploymentLabelsDefaultLabels = []string{"namespace", "deployment"}
 
-	deploymentMetricFamilies = []metric.FamilyGenerator{
+	deploymentMetricFamilies = []generator.FamilyGenerator{
 		{
 			Name: "kube_deployment_created",
 			Type: metric.Gauge,