--- a/internal/store/endpoint.go
+++ b/internal/store/endpoint.go
@@ -25,6 +25,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"k8s.io/kube-state-metrics/pkg/metric"
+	generator "k8s.io/kube-state-metrics/pkg/metric_generator"
 )
 
 var (
@@ -32,7 +33,7 @@ var (
 	descEndpointLabelsHelp          = "Kubernetes labels converted to Prometheus labels."
 	descEndpointLabelsDefaultLabels = []string{"namespace", "endpoint"}
 
-	endpointMetricFamilies = []metric.FamilyGenerator{
+	endpointMetricFamilies = []generator.FamilyGenerator{
 		{
 			Name: "kube_endpoint_info",
 			Type: metric.Gauge,