--- a/internal/store/hpa.go
+++ b/internal/store/hpa.go
@@ -27,6 +27,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"k8s.io/kube-state-metrics/pkg/metric"
+	generator "k8s.io/kube-state-metrics/pkg/metric_generator"
 )
 
 type MetricTargetType int
@@ -50,7 +51,7 @@ var (
 
 	targetMetricLabels = []string{"metric_name", "metric_target_type"}
 
-	hpaMetricFamilies = []metric.FamilyGenerator{
+	hpaMetricFamilies = []generator.FamilyGenerator{
 		{
 			Name: "kube_hpa_metadata_generation",
 			Type: metric.Gauge,