**What this PR does / why we need it**:

This PullRequest allows a third-party project to reuse the kube-state-metrics `collectors` by adding a way to configure a custom `cache.Store` generator function in the `store.Builder`.

In summary, the modifications are:
* Use `cache.Store` interface in “Builder”. It is a lot of code changes, but it is mainly a one to one replacement.
* Remove the FamilyByteSlicer interface and provide access to “metrics.Family” directly. Also, I have to split into two packages “metrics” package (creation of “metrics generator”) package to solve a cyclic import issue.
* To keep all the logic in the internal package, I created a public package that implements a kind of “bridge” pattern and also an interface for the Builder.


**Which issue(s) this PR fixes**

Fixes #973