@brancz 2019-12-05
hmm it looks like this unfortunately has a rather big hit on our benchmarks (see the benchmark job on CI) .. 10% increase in /metrics response is something I'm not entirely sure we want.

cc @tariq1890 @LiliC what do you think?