@LiliC 2019-12-06
Anything that impacts the performance in a negative way is not something we want, while I understand it would be good to have this. 10% is a lot IMO.