@brancz 2019-12-14
Generally this looks fine, but I'm still a little confused about the consumption model. Could you draft a bit of code showing what you're intending to do?