@clamoriniere 2019-12-17
Hi @brancz 
sorry for the late reply.

to illustrate the usage of this modification, I built this POC: https://github.com/clamoriniere/ddksm
some key entries in the code:
- the custom metrics store that stores `metric.Family` by resources and generates Datadog metrics: https://github.com/clamoriniere/ddksm/blob/master/pkg/store/store.go
- the new `builder` method `WithCustomGenerateStoreFunc()` is used in addition to the common `kube-state-metrics` configuration: https://github.com/clamoriniere/ddksm/blob/master/main.go#L125
- to trigger metrics generation and export to the `Dogstatd` server, the main process is looping over the `metricStore` and call the`.push()` method: https://github.com/clamoriniere/ddksm/blob/master/main.go#L133